refactor: modularize legal discovery interface
