+++ apps/legal_discovery/AGENTS.md
@@ -733,3 +733,8 @@ pip install python-dotenv flask gunicorn pillow requests neuro-san pyvis
 - Enabled chat-driven timeline updates with cross-links to depositions, exhibits and theories, plus summary endpoint and blur styling.
 - Next: broaden natural language date parsing and display linked events in dashboard.
 
+
+## Update 2025-08-09T15:00Z
+- Refactored Flask app into feature blueprints for theories, binder and chat with feature toggles.
+- Updated chat UI and routing tests to use blueprint paths.
+- Next: expand binder capabilities and modularise remaining routes.
