+++ apps/legal_discovery/chat_state.py
@@ -1,0 +1,5 @@
+import queue
+
+user_input_queue = queue.Queue()
+
+__all__ = ["user_input_queue"]
