+++ apps/legal_discovery/extensions.py
@@ -1,0 +1,5 @@
+from flask_socketio import SocketIO
+
+socketio = SocketIO()
+
+__all__ = ["socketio"]
